Template for the website, animated intro and chat functionality
Just a template on the layout of what the website may look like, there's also an animated intro page before being directed to the main page. The chat functionality for the main page has begun. Fixed some window resizing issues such as the login boxes not resizing with the window, but now they scale completely to the window size.